Preliminary fix for issue #219

+++ pmm-tests/drop_databases.sh
@@ -9,15 +9,30 @@ MYSQL_USER=root
 
 if [[ "${CLIENT_NAME}" == "ps" ]]; then
   BASEDIR=$(ls -1td ?ercona-?erver-5.* | grep -v ".tar" | head -n1)
-  BASEDIR="$WORKDIR/$BASEDIR"
-
+  if [[ -z "$BASEDIR" ]]; then
+    echo "Started fresh run!"
+  else
+    BASEDIR="$WORKDIR/$BASEDIR"
+  fi
 elif [[ "${CLIENT_NAME}" == "ms" ]]; then
   BASEDIR=$(ls -1td mysql-5.* | grep -v ".tar" | head -n1)
-  BASEDIR="$WORKDIR/$BASEDIR"
+  if [[ -z "$BASEDIR" ]]; then
+    echo "Started fresh run!"
+  else
+    BASEDIR="$WORKDIR/$BASEDIR"
+  fi
 
 elif [[ "${CLIENT_NAME}" == "pxc" ]]; then
   BASEDIR=$(ls -1td Percona-XtraDB-Cluster-5.* | grep -v ".tar" | head -n1)
-  BASEDIR="$WORKDIR/$BASEDIR"
+  if [[ -z "$BASEDIR" ]]; then
+    echo "Started fresh run!"
+  else
+    BASEDIR="$WORKDIR/$BASEDIR"
+  fi
 fi
 
-${BASEDIR}/bin/mysql --user=${MYSQL_USER} --socket=${MYSQL_SOCK} -e "drop database $DATABASE"
+if [[ -z "$BASEDIR" ]]; then
+  echo "Started fresh run!"
+else
+  ${BASEDIR}/bin/mysql --user=${MYSQL_USER} --socket=${MYSQL_SOCK} -e "drop database $DATABASE"
+fi
